+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## [1.0.1] -
+## [1.0.2] - 2022-04-29
+### Changed
+- Allow for dangling semicolons at end
+
+## [1.0.1] - 2022-04-16
 ### Changed
 - Allow for unknown content-disposition types and added a flag to enforce only known.